pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["RamonOrtisz <www.gamesetutoriaiss@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.3"
factory-boy = "^3.3.1"
django = "^5.1.3"
django-rest-framework = "^0.1.0"
pytest-django = "^4.9.0"
faker = "^33.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
